Fixed makefile compilation, now compiling with flags

+++ ex00/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: dbelpaum <dbelpaum@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 11:03:36 by dbelpaum          #+#    #+#             */
-/*   Updated: 2022/10/18 11:26:47 by dbelpaum         ###   ########.fr       */
+/*   Updated: 2022/10/24 10:27:05 by dbelpaum         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	main(int argc, char const *argv[])
 	{
 		for (int i = 1; i < argc; i++)
 		{
-			for (int j = 0; j < strlen(argv[i]); j++)
+			for (size_t j = 0; j < strlen(argv[i]); j++)
 			{
 				std::cout << (char)toupper(argv[i][j]);
 			}
